Move CreateAccount to sub-package 'ui'
Centralise all UI elements in one package ('com.ui').
